requirements.txt

# external requirements
click~=8.1.3
Sphinx
coverage
#awscli
flake8
python-dotenv>=0.5.1
hydra-core
omegaconf
tensorboard

# backwards compatibility
pathlib2

torch~=2.0.0
numpy
torchtext~=0.15.1
setuptools~=49.2.1

matplotlib==3.7.1
numpy==1.23.5
PyYAML==6.0
scikit_learn==1.2.2
setuptools

#for build_features.py (added by Maude)
portalocker>=2.0.0
torchtext
protobuf< 5.0.0 